requirements.txt
inference-gpu
supervision
opencv-python
streamlit>=1.35.0
numpy
Pillow